+++ src/algo_royale/clients/db/dao/trade_dao.py
@@ -1,6 +1,5 @@
 # src/algo_royale/clients/db/dao/trade_dao.py
 from datetime import datetime
-from decimal import Decimal
 from uuid import UUID
 
 import psycopg2
@@ -59,8 +58,8 @@ class TradeDAO(BaseDAO):
         symbol: str,
         action: str,
         settlement_date: datetime,
-        price: Decimal,
-        quantity: int,
+        price: float,
+        quantity: float,
         executed_at: datetime,
         order_id: UUID,
         user_id: str,
